fix(configuration): add new createEfConfig & update api element
