+++ packages/elements/src/configuration/index.ts
@@ -1,2 +1,3 @@
 export * from './elements/configuration.js';
+export * from './helpers/configuration.js';
 export { efConfig, type Config } from './helpers/context.js';
